extract ChildNodes component to its own file
